+++ src/providers/twitch/api/Helix.hpp
@@ -300,6 +300,16 @@ struct HelixChannelEmote {
     }
 };
 
+enum class HelixAnnouncementColor {
+    Blue,
+    Green,
+    Orange,
+    Purple,
+
+    // this is the executor's chat color
+    Primary,
+};
+
 enum class HelixClipError {
     Unknown,
     ClipsDisabled,
@@ -335,6 +345,14 @@ enum class HelixDeleteChatMessagesError {
     UserNotAuthenticated,
     UserNotAuthorized,
     MessageUnavailable,
+
+    // The error message is forwarded directly from the Twitch API
+    Forwarded,
+};
+
+enum class HelixSendChatAnnouncementError {
+    Unknown,
+    UserMissingScope,
 
     // The error message is forwarded directly from the Twitch API
     Forwarded,
@@ -511,6 +529,13 @@ public:
         FailureCallback<HelixRemoveChannelModeratorError, QString>
             failureCallback) = 0;
 
+    // https://dev.twitch.tv/docs/api/reference#send-chat-announcement
+    virtual void sendChatAnnouncement(
+        QString broadcasterID, QString moderatorID, QString message,
+        HelixAnnouncementColor color, ResultCallback<> successCallback,
+        FailureCallback<HelixSendChatAnnouncementError, QString>
+            failureCallback) = 0;
+
     virtual void update(QString clientId, QString oauthToken) = 0;
 };
 
@@ -650,6 +675,13 @@ public:
     void removeChannelModerator(
         QString broadcasterID, QString userID, ResultCallback<> successCallback,
         FailureCallback<HelixRemoveChannelModeratorError, QString>
+            failureCallback) final;
+
+    // https://dev.twitch.tv/docs/api/reference#send-chat-announcement
+    void sendChatAnnouncement(
+        QString broadcasterID, QString moderatorID, QString message,
+        HelixAnnouncementColor color, ResultCallback<> successCallback,
+        FailureCallback<HelixSendChatAnnouncementError, QString>
             failureCallback) final;
 
     void update(QString clientId, QString oauthToken) final;